Clean up operator code
