+++ src/Web/ClientApp/src/components/inspection/Details.tsx
@@ -8,8 +8,7 @@ import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
 import KeyboardArrowDownIcon from '@material-ui/icons/KeyboardArrowDown';
 import KeyboardArrowUpIcon from '@material-ui/icons/KeyboardArrowUp';
 import { useInputTypes } from './Types';
-import { InspectionSheet, Equipment, InspectionItem } from '../../entities';
-import { initialState } from '../../use-cases/InspectionSheetInteractor';
+import { InspectionSheet, Equipment, InspectionItem, InspectionSheetInitialState } from '../../entities';
 import { InspectionGroup, InspectionType } from '../../typescript-fetch';
 
 const useStyles = makeStyles((theme: Theme) =>
@@ -76,7 +75,7 @@ const Row: FC<RowProps> = ({ equipment }): JSX.Element => {
 
 export const Details = ({ match }: any): JSX.Element => {
   const sheetId = match.params.id;
-  const [inspectionSheet, setInspectionSheet] = useState<InspectionSheet>(initialState());
+  const [inspectionSheet, setInspectionSheet] = useState<InspectionSheet>(InspectionSheetInitialState);
   const [groups, setGroups] = useState<InspectionGroup[]>([]);
   const [types, setTypes] = useState<InspectionType[]>([]);
