+++ src/Web/ClientApp/src/use-cases/index.ts
@@ -1,3 +1,2 @@
-export * from './InspectionItemOperator';
 export * from './InspectionItemInteractor';
 export * from './InspectionSheetInteractor';
